requirements.txt
graphlib
threaded
paho-mqtt
tabulate
pandas
numpy